Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> An alternative approach would be to leave TARGET_CPU_VARIANT for mainline devices, and set a TARGET_GENERIC_CPU_VARIANT := true.  Then in envsetup.mk after reading the BoardConfig.mk stash away the original values to use with the properties and overwrite them with generic.

Yeah, that seems reasonable too. One problem I have with that is that it may catch people by surprise that TARGET_CPU_VARIANT doesn't behave like it used to. I think I'd like to keep the TARGET_CPU_VARIANT refer to static compilation setting, but it's a little hard to come up with something cohesive/logical until the support is improved to support multiple variants for different partitions. We may be able to have something like this later:

TARGET_CPU_VARIANT := <specific> (this will be the default)
TARGET_CPU_VARIANT_SYSTEM := generic (override on a per-partition basis)

I'll submit this as-is but I forsee refactoring of this in my future .. :)

Patch-set: 1
Reviewer: Gerrit User 1060959 <1060959@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0