Remove a few entries from PRODUCT_HOST_PACKAGES
adbd shouldn't be needed on the host in most cases

selinux_policy_system is a device PHONY_PACKAGE that will stop working
soon in PRODUCT_HOST_PACKAGES. I talked to jeffv and he's fine removing
it for now, adding back individual binaries if desired, or creating a
new host PHONY_PACKAGE.

shell_and_utilities_system is a similar device phony module. enh is fine
with it not being installed on the host by default for now, and we can
add them back as needed.

Test: treehugger
Change-Id: I1938f9f4ea5dbea2905baaf301473b617de9c530