Remove 'column' from modules target
The standard target modules stopped working when restrictions in
soong were introduced to which binaries are allowed to be used during
builds.

Listing a very large amount of modules in columns does not make it
more readable and harder to work with in tools so just drop it.

Bug: 129800175
Test: make modules

(cherry picked from commit 7005d4f4ccc022c102217511324295c0997f88f9)

Change-Id: I317f6eb4ba5ea633264939a9a58f716bdb36596e