Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Silently ignoring notice files seems dangerous, and a notice file of a module bundled in an apex module still needs to be tracked somehow if the apex module is installed.

What error case do we try to prevent with this? For example, if a 3rd party module with a NOTICE file is marked as uninstallable, this generates an error. What do we wanna do in this case? Do we delete or rename the file just to make this rule happy?

The reason why I mentioned apex is because this is a dormant bug for apex components. It's been going unnoticed because all uninstallable apex components at the moment either don't have a notice file in the same dir or have one with a different name than NOTICE. Create a NOTICE file under system/timezone/output_data, then the build fails because of this line. Notice file tracking for apex modules are being done separately in Soong with or without this change.

Patch-set: 3
Reviewer: Gerrit User 1060667 <1060667@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0