Use --ignore_missing_files when creating symbols.zip
symbols.zip is created while files that are not part of the platform
build but were requested on the Make command line (for example by
passing "tests") are still being installed.  Installation often
involves removing and then recreating the file.  If the file
exists when the list of files to be zipped is created, but is
removed when soong_zip is zipping, it fails with an error.  Pass
--ignore_missing_files to silence the error and keep zipping.

Fixes: 79503862
Test: m checkbuild
Change-Id: I1fbbf1f1396948288ee133c319b2e19dee97983b