Update patch set 1
Patch Set 1:

Symlinks should not be part of the dependency graph (at least for now).  You can use this rule to create a symlink, but you need to depend on the target of the symlink, and not the symlink itself, if you want to rebuild when it changes.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>