temporarily disable vendor file check for mainline
We want to use prebuilts for apex modules for qt-dev, but it is
running afoul of the vendor file check.  Disable the check for
now until we figure out a better solution.

Using Merged-In instead of DNM

Bug: 137033385
Test: Forrest run build_test
Change-Id: I9db5cb227780ede6aaff0070cd2fd59e95e635e6
Merged-In: 874b7a7766a569613dcd3ae526eaa6e1d4b78866
Merged-In: I99431a9a342e9b0617510e250597f3024ef39322