Update patch set 2
Patch Set 2:

This will break Mac builds when paired with the previous change. Python3 is not guaranteed to be available on Mac (or guaranteed to be any specific Python3 version on any platform). We're aiming to get this fixed this quarter so that we've got a hermetic version that can be used.

Patch-set: 2
CC: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Work-in-progress: false