fix: set built_module in misc_prebuilt_internal.mk
*_prebuilt_internal.mk should set "built_module" with LOCAL_BUILT_MODULE
on success, since it is used to set dependency between built module and
additional dependencies (LOCAL_ADDITIONAL_DEPENDENCIES).

In aosp/970424, it didn't set it properly.

As a consequence, setting additional dependencies to a built module was
wrong for various classes of modules(ETC, SCRIPT, DATA)

For example, prebuilt_etc_xml module(com.google.vr.platform.xml) has
an additional depdendency(timestamp of xmllint). But the dependency is
set to unexected module(com.android.support.apexer.soong_zip).
This is because that
1) built_module is not set in misc_prebuilt_internal.mk
2) built_module is not reset after use and soong_zip was last module
which set built_module variable.

Bug: 132925121
Test: m com.google.vr.platform.xml && check build-<TARGET>.ninja to see
if there is additional dependency as described above

Change-Id: I74fdf6b6ad583f3c08dc0b8fc8595d7ef9cc7306