Remove libcameraservice and libcamera_client from base_system
Libcameraservice is only used by cameraserver, and is explicitly listed
as a dependency.  Libcamera_client is used by multiple places, but each
of them lists it as a depedency as well, so it's not needed here.

Removing the unused 64-bit version of libcameraservice will save ~2 MB
on the system partition.

Bug: 138403869
Test: atest CtsCameraTestCases
Merged-In: I196f869350900e7cc1521bc397c6ecff28decb6f
Change-Id: I196f869350900e7cc1521bc397c6ecff28decb6f