Add misc_info.txt needed to build super.img
Previously, misc_info.txt is included only when super.img is included
in the update package. This CL adds it to the update package so
Goldfish-based products for emulator can use it to create a new super.img
later in TF pipeline so mixed-image tests can be performed.

A typical example for such mixed-image tests is GSI (aosp_x86) on
sdk_gphone_x86 or older aosp_x86 vendor images.

Bug: 138545354
Test: $ lunch aosp_x86-userdebug; m - j
      $ unzip -l $OUT/aosp*.zip | grep misc
      526  2019-07-22 14:50   misc_info.txt

Change-Id: I111716f499b031c03f2896299b7436a7611841b8