Add libneuralnetworks_packageinfo to dependencies.
libbinder communication from APEX is not allowed, because it doesn't
form a stable way to communicate. After an update, if system libbinder
and APEX libbinder are out of sync, then they may fail to work with each
other.

libbiner usage is moved into libneuralnetworks_packageinfo library,
which is a stable API library.

Test: build & flash crosshatch
Bug: 139282353
Change-Id: I7238025af13e9430c266a1aea69f0d0381110df8