Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > If the build.prop contains multiple lines which have the same key but with the different values, such as "ro.build.fingerprint=abc" and "ro.build.fingerprint=xyz", the "LoadDictionaryFromLines" save the value with "xyz" for the key "ro.build.fingerprint".
> > > 
> > > But the system in the phone get the value is "abc", that is the problem.
> > 
> > This is very serious in some cases with the ota update.
> 
> 1. Can you please let us know how this fix can avoid such case? This function usually handles content from a single file, so this change may not avoid the case you faced.
> 
> 2. I'm not sure if we should filter out ro. property collision from this function. I'm not quite sure that in every case this function will handle system properties first, and also in every case system image should have priority. Can you please handle this from more specific function, such as ota_from_target_files.py if it is the tool you've used?

1. Ｉ can give a very typical example. Suppose there are two target versions，we call them as 'version A' and 'version B'. I need to use these two versions to do differential package upgrades by run the tool 'ota_from_target_files.py'. The 'system/build.prop' contains two lines which have the same key 'ro.build.fingerprint', but with the different values in the 'vension A'. As shown below:

...(some other lines)...
ro.build.fingerprint=AAA
...(some other lines)...
ro.build.fingerprint=aaa
...(some other lines)...

And the 'version B' also has similar features. As shown below:

...(some other lines)...
ro.build.fingerprint=BBB
...(some other lines)...
ro.build.fingerprint=bbb
...(some other lines)...

The source version is 'version A', and the target version is 'version B'. When I make a differential package with the tool 'ota_from_target_files.py', it will generate the file 'updater-script', in the file 'updater-script' contains the following content:

getprop("ro.build.fingerprint") == "aaa" ||
    getprop("ro.build.fingerprint") == "bbb" ||
    abort("E3001: Package expects build fingerprint of aaa or bbb; this device has " + getprop("ro.build.fingerprint") + ".");

But the property 'ro.build.fingerprint' on the phone which run the rom 'version A' is 'AAA', and the property 'ro.build.fingerprint' on the phone which run the rom 'version B' is 'BBB', that is the problem. And the upgrade with the 'OTA' failed.

The cause of the above problem is that the properties with 'ro.' are actually "write-once". So the "ro.*" properties in the 'build.prop' should be only loaded once in the tool 'ota_from_target_files.py' to keep the same behavior as the actual phone operation.


2. Your worry is justified. I will handle it in another function, but not in the 'ota_from_target_files.py'. Because the dictionary has been processed in the 'common.py', and it can not be processed normally in the 'ota_from_target_files.py'. I will submit a new patch.

Patch-set: 1