Update patch set 2
Patch Set 2:

> Thanks for pointing to the code. But I still think the behavior of how ro.property is initialized is not a guaranteed behavior. e.g. init could switch to read the build.prop in a map (which overrides) and then write once. 

Init actually does do this.  We switched the behavior in Q.

Since the duplication in the build.prop almost always indicates an error in the build system, how about adding an argument (e.g. no_dup) to optionally assert and error out in LoadDictionaryFromLines? Then if there is an error, we can go back and fix the build system. What do you think.

I 100% agree.  The build should never output the same property in a single build.prop file.  If it is doing this, that needs to be fixed.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>