Update patch set 4
Patch Set 4:

This is likely to cause lots of product-specific breakages any time a module is changed that is installed to multiple products, some with the flag and some without.  Is there a plan to relatively quickly get to a state where this is turned on everywhere?  Would this be better using the BUILD_BROKEN_* style to blacklist failing products?

Patch-set: 4
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>