base_system.mk: Switch to wifi mainline apex
Also, explicitly include the cacerts_wfa in builds. This used to be
included via the "required" tag in wifi-service build rule. We probably don't
want to deliver CA certificates to OEM devices via mainline updates. So,
leaving it out of the wifi mainline module & explicitly including it in
the base makefile.

Bug: 144722612
Test: Device boots up & connects to wifi networks.
Change-Id: I3e4487f9fad765e442bf8335ec1f21de3a47674d