build: enable Perfetto by default
Perfetto has been enabled by default on Pixel devices since P and has
proved to be stable and helped find performance issues on both P and Q.

On R+, iorapd has a dependency on Perfetto and thus we need to be enabled
by default.

Note that CTS tests have ensured that any downstream consumers of Android
have not accidentally broken Perfetto in old releases so this should be
safe to introduce and maintain going forward.

The two daemons of Perfetto (traced and traced_probes) are located on the
/system partition so for consistency, we have chosen to also enable the
build property on the system partition.

Googlers can see the below link for more context and detailed info.

Change-Id: I00fa7d4c97eda4d0ab83520d0f428bca83707d8a
Context (Googlers only): go/perfetto-by-default
Bug: 137963570
Test: checking product/build.prop in Treehugger builds