Make qt-qpr1-dev-plus-aosp use AOSP.MASTER as BUILD_ID.
In ag/7020561, after branching of qt-dev, the BUILD_ID was changed to
"QT" and the change was tagged as "DO NOT MERGE" to prevent it flowing
into master.  However, because the -plus-aosp branches was using the
same API level and "DO NOT MERGE" will not be respected there, it would
eventually be merged to stage-aosp-master and promoted to AOSP master.

Address this by using AOSP.MASTER which will cause a merge conflict when
the merge happens.

Bug: 145935472
Change-Id: Iaa17a2c29f1f3cf28f28761a18b0d480b0253fac
Merged-In: I5c28225d61996cb719ae384d7b0828776e84f8d5