Initialize common_proguard_flags for each module
Id938aa35e87c9c69769a3184dfea06fd02299688 removed the initial
assignment to common_proguard_flags, which both makes it grow as
each module is parsed and also turns it into a lazily evaluated
variable.  That causes the flags passed to R8 to be the flags
computed for the current module multiplied by the number of previously
parsed modules.

Fixes: 146150575
Test: m checkbuild
Change-Id: I87c9625496f482896be6867124f2f12ff662cd57