Add boringssl_self_test to PRODUCT_PACKAGES
boringssl_self_test is a required module of the conscrypt.jar.
Therefore, it was installed along with the conscrypt.jar. However,
conscrypt.jar is no longer available for the platform but only for the
conscrypt APEX. The problem here is that the conscrypt APEX having
conscrypt.jar doesn't carry its 'required' property. As a result,
boringssl_self_test is missing. This is b/146549048.

Until the bug is fixed, manually adding the test binary to
PRODUCT_PACKAGES to ensure that it is installed always.

Bug: 146549048
Test: m, device boots

Change-Id: I72dbc51c267f8d8683b70c07ae726c3a3eba0aea