Compare existing ABI dump list with LSDUMP_PATHS
This commit fixes the check for redundant files in prebuilts/abi-dumps.
Some ABI dump file names are different from the module names. The list
of existing files should be compared with LSDUMP_PATHS which are the
files generated by soong.

This commit also adds the check for the files in
prebuilts/abi-dumps/platform.

Bug: 147409497
Test: make
Change-Id: If304afb116e9b5d3cb7ceaf74822d5a19ebe1a35