Remove keymaster libraries from VNDK
These keymaster libraries constitute a service to vendors
who want/need to implement a keymaster trusted app, or want
to wrap an older keymaster HAL when updating an older device
to a newer Android version. It is not necessary to implement
or link against these libraries to be Android or Treble
compliant. Many vendors have completely independent
implementations of keymaster. Vendors should be able to
provide there own versions of these libraries or ship without
them entirely. /system does not depend on /vendor to have
a compatible version of these libraries.

Test: N/A
Bug: 139955894
Change-Id: I9286b36d26f09e26763df17758f9bec9e2b60ec8