Update patch set 3
Patch Set 3: Code-Review+2

This tool is relatively small (~1.3 MB) so this is fine for now.  If we do need it in the future, we can store a copy of this tool, once per toolchain update, in a location accessible from tradefed.  (We already have a script that uploads prebuilts to gerrit so this'd be automated as part of that).

Patch-set: 3
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2