Update patch set 4
Patch Set 4:

(1 comment)

so the point of this change is to make system_ext reusable across products? do we think that's actually achievable? I wonder if we should change the syntax of '*' to make it clearer, instead of changing the meaning of it

Patch-set: 4
Reviewer: Gerrit User 1060959 <1060959@85c56323-6fa9-3386-8a01-6480fb634889>