strings.mk: Allow empty RHS values for inputs of collapse-pairs
This CL is to make the function collapse-pairs(), which is applied to
lists of "property=value" pairs, allow empty properties (i.e. lines of
the form "empty.prop=").
Currently, the function only allows empty properties at the end (see
https://android-review.googlesource.com/c/platform/build/+/722908).
So empty properties in the middle result in broken property files (e.g.
"a=b c= d=e" is transformed to "a=b c=d=e", in which the property "c",
which originally has no value, is interpreted as having the value "d=e",
whereas the property "d", which originally has the value "e", is missing
from the resulting property file).
This CL revises the function so as to keep empty properties in the
middle as is (e.g. it returns "a=b c= d=e" for the above example), while
preserving the behavior for well-formed lists like "a=b c= d e = f".

Bug: 152379493
Test: make
Change-Id: I35faeaedc3bc42e56e01201baf7ea6805a610439