Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Different targets treat this as an error or not. I'm surprised to only see the "generic" target fail, and not other targets on ab/ (unless they just haven't shown up yet) +Anton
> 
> (This should fail if you do `lunch generic-eng; m nothing` locally)

This check is only enabled for generic because a lot of products (like pixels...) that inherit the other mks (*_system.mk) all violate this rule. Enabling it for generic ensures that the generic *_system.mk files stay in a good place. If we cleanup the pixels, we could enable it in e.g. mainline_system.mk too.

Patch-set: 1