Add boringssl_self_test to base PRODUCT_PACKAGES.
When building from source this gets installed as a requirement
from Conscrypt, but needs to be present when Conscrypt is
shipped as a pre-built too.

I think the correct solution here is actually to make libcrypto
rather than Conscrypt require boringssl_self_test but Soong
considers that a circular dependency.

Bug: 155520949
Test: m installclean && m

Change-Id: I77b2921d1d7b6be493cd581c46adbadd1f186fb0