Remove /default.prop
It was kept for non-Treble devices, but now /system/build.prop is used
instead by having both system-configured sysprops and vendor-configured
sysprops. To summarize:

* For Treble compliant devices
- /system/build.prop has
  ADDITIONAL_DEFAULT_PROPERTIES
  PRODUCT_SYSTEM_DEFAULT_PROPERTIES
  props generated from buildinfo_common.sh
  props generated from buildinfo.sh
  props read from TARGET_SYSTEM_PROP file
  ADDITIONAL_BUILD_PROPERTIES

- /vendor/build.prop has
  PRODUCT_DEFAULT_PROPERTY_OVERRIDES
  props generated from buildinfo_common.sh
  props read from TARGET_VENDOR_PROP
  PRODUCT_PROPERTY_OVERRIDES

* For non-Treble devices,
PRODUCT_DEFAULT_PROPERTY_OVERRIDES moves from /vendor/build.prop to
/system/build.prop (just after PRODUCT_SYSTEM_DEFAULT_PROPERTIES).

Other props in /vendor/build.prop remain there but they are not used
because it's a symlink for /system/vendor/build.prop and init does not
follow symlink. Especially, PRODUCT_PROPERTY_OVERRIDES is ignored.
Someone might think this as a bug, which actually is, but this change
doesn't attempt to fix that, because 1) that's out of scope of this
change, and 2) we will anyway merge PRODUCT_PROPERTY_OVERRIDES and
PRODUCT_DEFAULT_PROPERTY_OVERRIDES into one.

Bug: 117892318
Test: TH passes

Change-Id: I867f3615ef8c06650ab2ee96642bac4f6d8ca820