Use NetworkStack on system image
Now that NetworkStack was bumped to API 30, use NetworkStack instead of
NetworkStackNext on the system image.

This change should not be merged in branches that are used for building
S system images for internal testing. Note that the merged-in change
is not related to this change; it is just used to prevent this change
from merging into the aosp/master merge path.

Bug: 143586229
Test: m
Merged-In: Ibc74edb88904f953a1e5e6ceef44c461f381c9cf
Change-Id: I484e148fab80a01d2b20ebf48604d9e923fe7cd3