Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 4:
> > 
> > (2 comments)
> > 
> > I'm not sure why the additional changes are needed? Just calling module-load-list-copy-paths() (as it is currently) worked for me, producing a modules.load.recovery with the same order as what I had specified in BOARD_VENDOR_RAMDISK_RECOVERY_KERNEL_MODULES_LOAD .
> 
> Not sure you saw my above comment... The BOARD_*_KERNEL_MODULES_LOAD macros were intended to take full module paths from the build:

Well, that sucks, because no one is doing that in their board configurations, and because the only load order that truly makes sense and is easy to produce is the one coming from the modules.load file. This is generated by the kernel build. Except for this macro fragment, all the other macros just work just fine because they have separate args for module lists and another for modules load content. Module load content works in the other macros as a pure list of just module names, no paths even, because they get basename'd (or notdir'd) in practice.

Trying to convert the content in modules.load (which all the board configuration files are doing right now) to paths is 'ucky' and if we have to do that in the board configuration files its going to have a high cost.

> 
> https://source.android.com/preview/devices/architecture/kernel/kernel-module-support#first-stage-init

(404)?

> 
> I can see having the load list depend on all the modules instead of just the ones in the list, but I'm  not following why the changes on 772/773 (PS5) are needed?

Because the existing Board configuration files (which I used as my template w/o docs) are not doing the *_LOAD as you specified. And everything else just seems to work anyways.

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0