Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > This will force any build that wants coverage to build every java library defined in the tree, not just the ones necessary for the product.
> > 
> > That's not strictly true. It will only build every java library that is instrumented and not every java library is instrumented. There are also existing mechanisms to filter out the modules that can be instrumented, e.g. JavaCoveragePaths and JavaCoverageExcludePaths. e.g. When I build this on AOSP the generated jacoco-report-classes-all.jar contains 71 files.
> > 
> > Any suggestions for how the list of jacoco jars could be filtered to further restrict it to the set of ? Or is there a better approach altogether?
> 
> The command I used to generate the jacoco-report-classes-all.jar is:
> build/soong/soong_ui.bash --make-mode EMMA_INSTRUMENT=true EMMA_INSTRUMENT_FRAMEWORK=true JAVA_COVERAGE_PATHS="*" out/target/product/taimen/jacoco-report-classes-all.jar
> 
> So, AFAIK that will ensure that every possible java library that could be instrumented is instrumented.

For comparison a previously generated jacoco-report-classes-all.jar contained 900+ files and was ~4G in size. For a similar build configuration this change generates one containing ~120 files and was ~200M in size. The improved compression ratio is most likely because those 120 files contain a number of duplicates (e.g. one copy of each library for each APEX that contains it).

While not all libraries are instrumented it seems as though the main difference in size is due to the fact that the code for handling LOCAL_SOONG_JACOCO_REPORT_CLASSES_JAR is duplicated in core/soong_app_prebuilt.mk but I did not update that to match what I did for the java library code.

I am not sure what purpose is served by adding app files to jacoco-report-classes-all.jar.

Patch-set: 1
Label: Presubmit-Ready=0