Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> We've had minimal targets before that pretty quickly stopped working. If we're going to add one of these generic/reusable makefile we need to have clear guidance for when something belongs in this target. Is that "never", unless your new lunch target stops booting to shell successfully? If so, how is that going to be monitored?

The guidance could be:
If adding to base_system.mk, consider if the component is needed to boot to shell. If so, add it here instead.

If something does not fit to base_system.mk, but is required to boot to shell, it should go to the launch target.

There will be some monitoring needed. We intend to have CI that will monitor this and fix any issues, but this is not publicly viewable.
Also Peter Collingbourne is currently using fvp-eng. He might have another idea...

Patch-set: 1